+++ platform/darwin/src/NSDate+MGLAdditions.h
@@ -5,6 +5,9 @@
 @interface NSDate (MGLAdditions)
 
 /// Converts from a duration in seconds to a duration object usable in mbgl.
-mbgl::Duration MGLDurationInSeconds(NSTimeInterval duration);
+mbgl::Duration MGLDurationInSecondsFromTimeInterval(NSTimeInterval duration);
+
+/// Converts from an mbgl duration object to a duration in seconds.
+NSTimeInterval MGLTimeIntervalFromDurationInSeconds(mbgl::Duration duration);
 
 @end