+++ platform/darwin/src/NSDate+MGLAdditions.mm
@@ -2,9 +2,14 @@
 
 @implementation NSDate (MGLAdditions)
 
-mbgl::Duration MGLDurationInSeconds(NSTimeInterval duration)
+mbgl::Duration MGLDurationInSecondsFromTimeInterval(NSTimeInterval duration)
 {
     return std::chrono::duration_cast<mbgl::Duration>(std::chrono::duration<NSTimeInterval>(duration));
+}
+
+NSTimeInterval MGLTimeIntervalFromDurationInSeconds(mbgl::Duration duration)
+{
+    return duration.count();
 }
 
 @end