+++ platform/macos/app/MapDocument.m
@@ -683,6 +683,9 @@ NS_ARRAY_OF(id <MGLAnnotation>) *MBXFlattenedShapes(NS_ARRAY_OF(id <MGLAnnotatio
 }
 
 - (IBAction)manipulateStyle:(id)sender {
+    self.mapView.style.transitionDuration = 5;
+    self.mapView.style.transitionDelay = 1;
+
     MGLFillStyleLayer *fillStyleLayer = (MGLFillStyleLayer *)[self.mapView.style layerWithIdentifier:@"water"];
     
     MGLStyleValue *colorFunction = [MGLStyleValue<NSColor *> valueWithStops:@{